+++ pkg/jwt/logger.go
@@ -4,4 +4,5 @@ import "context"
 
 type logger interface {
 	ErrUserAuth(context.Context, error) error
+	InfoGeneric(context.Context, string) error
 }