+++ pkg/jwt/test/logger.go
@@ -9,6 +9,9 @@ type Logger struct{}
 
 func (Logger) ErrUserAuth(ctx context.Context, err error) error {
 	fmt.Println(err)
-
+	return nil
+}
+func (Logger) InfoGeneric(ctx context.Context, msg string) error {
+	fmt.Println(msg)
 	return nil
 }